requirements.txt
anyio==3.7.1
asks==3.0.0
async-generator==1.10
asyncclick==8.1.3.4
attrs==23.1.0
h11==0.14.0
idna==3.4
loguru==0.7.0
outcome==1.2.0
python-dotenv==1.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
